fixes #1988 Add maskConfigProperties to server.yml to control the module registry masking
