+++ server-config/pom.xml
@@ -25,38 +25,30 @@
         <relativePath>../pom.xml</relativePath>
     </parent>
 
-    <artifactId>utility</artifactId>
+    <artifactId>server-config</artifactId>
     <packaging>jar</packaging>
-    <description>A utility module that contains all the utilities shared across other modules.</description>
+    <description>A server config of server module.</description>
 
     <dependencies>
         <dependency>
             <groupId>com.networknt</groupId>
-            <artifactId>server-config</artifactId>
+            <artifactId>config</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>com.fasterxml.jackson.core</groupId>
+            <artifactId>jackson-databind</artifactId>
         </dependency>
         <dependency>
             <groupId>org.slf4j</groupId>
             <artifactId>slf4j-api</artifactId>
         </dependency>
         <dependency>
-            <groupId>commons-codec</groupId>
-            <artifactId>commons-codec</artifactId>
-        </dependency>
-
-        <dependency>
             <groupId>ch.qos.logback</groupId>
             <artifactId>logback-classic</artifactId>
-            <scope>test</scope>
         </dependency>
-
         <dependency>
             <groupId>junit</groupId>
             <artifactId>junit</artifactId>
-            <scope>test</scope>
-        </dependency>
-        <dependency>
-            <groupId>com.networknt</groupId>
-            <artifactId>config</artifactId>
             <scope>test</scope>
         </dependency>
     </dependencies>
