+++ server-config/src/test/java/com/networknt/server/ServerConfigTest.java
@@ -31,7 +31,7 @@ public class ServerConfigTest {
     public void testNullEnv() {
         config.clear();
         // ensure that env is null if it is missing in the server.yml
-        ServerConfig serverConfig = (ServerConfig)Config.getInstance().getJsonObjectConfig(CONFIG_NAME, ServerConfig.class);
+        ServerConfig serverConfig = ServerConfig.getInstance("server");
         Assert.assertNull(serverConfig.getEnvironment());
         Assert.assertEquals("petstore", serverConfig.getServiceName());
     }
@@ -39,8 +39,8 @@ public class ServerConfigTest {
     @Test
     public void testDefaultServerOptions() {
         config.clear();
-        ServerConfig serverConfig = (ServerConfig)Config.getInstance().getJsonObjectConfig(CONFIG_NAME, ServerConfig.class);
-        ServerOption.serverOptionInit(config.getInstance().getJsonMapConfigNoCache(CONFIG_NAME),serverConfig);
+        ServerConfig serverConfig = ServerConfig.getInstance("server");
+        ServerOption.serverOptionInit(serverConfig.getMappedConfig(),serverConfig);
         Assert.assertEquals(1024*16, serverConfig.getBufferSize());
         Assert.assertEquals(Runtime.getRuntime().availableProcessors() * 2, serverConfig.getIoThreads());
         Assert.assertEquals(200, serverConfig.getWorkerThreads());
@@ -52,8 +52,8 @@ public class ServerConfigTest {
     @Test
     public void testInvalidServerOptions() {
         config.clear();
-        ServerConfig serverConfig = (ServerConfig)Config.getInstance().getJsonObjectConfig("server_invalid_option", ServerConfig.class);
-        ServerOption.serverOptionInit(config.getInstance().getJsonMapConfigNoCache("server_invalid_option"), serverConfig);
+        ServerConfig serverConfig = ServerConfig.getInstance("server_invalid_option");
+        ServerOption.serverOptionInit(serverConfig.getMappedConfig(),serverConfig);
         Assert.assertEquals(1024*16, serverConfig.getBufferSize());
         Assert.assertEquals(Runtime.getRuntime().availableProcessors() * 2, serverConfig.getIoThreads());
         Assert.assertEquals(200, serverConfig.getWorkerThreads());
@@ -66,8 +66,8 @@ public class ServerConfigTest {
     @Test
     public void testValidServerOptions() {
         config.clear();
-        ServerConfig serverConfig = (ServerConfig)Config.getInstance().getJsonObjectConfig("server_valid_option", ServerConfig.class);
-        ServerOption.serverOptionInit(config.getInstance().getJsonMapConfigNoCache("server_valid_option"), serverConfig);
+        ServerConfig serverConfig = ServerConfig.getInstance("server_valid_option");
+        ServerOption.serverOptionInit(serverConfig.getMappedConfig(),serverConfig);
         Assert.assertEquals(10000, serverConfig.getBufferSize());
         Assert.assertEquals(1, serverConfig.getIoThreads());
         Assert.assertEquals(100, serverConfig.getWorkerThreads());
@@ -80,8 +80,8 @@ public class ServerConfigTest {
     @Test
     public void testMaxTransferFileSize() {
         config.clear();
-        ServerConfig serverConfig = (ServerConfig)Config.getInstance().getJsonObjectConfig(CONFIG_NAME, ServerConfig.class);
-        ServerOption.serverOptionInit(config.getInstance().getJsonMapConfigNoCache(CONFIG_NAME),serverConfig);
+        ServerConfig serverConfig = ServerConfig.getInstance();
+        ServerOption.serverOptionInit(serverConfig.getMappedConfig(), serverConfig);
         Assert.assertEquals(1000000, serverConfig.getMaxTransferFileSize());
     }
 }
