+++ server-config/src/test/resources/config/server_invalid_option.yml
@@ -1,0 +1,5 @@
+bufferSize: -200
+ioThreads: 0
+workerThreads: 0
+backlog: 0
+alwaysSetDate: false
