+++ server-config/src/test/resources/config/server_valid_option.yml
@@ -1,0 +1,7 @@
+bufferSize: 10000
+ioThreads: 1
+workerThreads: 100
+backlog: 10000
+alwaysSetDate: false
+serverString: TEST
+allowUnescapedCharactersInUrl: true
