+++ server/src/test/java/com/networknt/server/TestServer.java
@@ -39,7 +39,7 @@ public class TestServer extends ExternalResource {
     }
 
     public ServerConfig getServerConfig() {
-        return Server.getServerConfig();
+        return ServerConfig.getInstance();
     }
 
     @Override
